# PERF: Series.add for pyarrow string/binary dtypes