- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.1.0.rst` file if fixing a bug or adding a new feature.

```
import pandas as pd
import pandas._testing as tm
import pyarrow as pa

N = 1_000_000
s1 = pd.Series(tm.makeStringIndex(N), dtype=pd.ArrowDtype(pa.string()))
s2 = pd.Series(tm.makeStringIndex(N), dtype=pd.ArrowDtype(pa.string()))

%timeit s1 + s2
# 136 ms ± 3.25 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)   -> main
# 61 ms ± 2.86 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)    -> PR

%timeit s1 + "abc"
# 200 ms ± 12.2 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)   -> main
# 64.5 ms ± 2.95 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  -> PR
```